requirement.txt
pycryptodomex==3.20.0
pypiwin32==223
pywin32==306
